OSX compile error fix

+++ memcached.c
@@ -8110,7 +8110,7 @@ static void detlongq_write(char client_ip[], char* command, int cmdnum)
     ptm = localtime(&val.tv_sec);
 
     snprintf(inputstr, DETECT_INPUT_SIZE, "%02d:%02d:%02d.%06ld %s %s\n",
-                ptm->tm_hour, ptm->tm_min, ptm->tm_sec, val.tv_usec, client_ip, command);
+                ptm->tm_hour, ptm->tm_min, ptm->tm_sec, (long)val.tv_usec, client_ip, command);
     inputlen = strlen(inputstr);
 
     memcpy(detlongq.buffer[cmdnum].data + detlongq.buffer[cmdnum].offset, inputstr, inputlen);
